api-builder: move ngdoc stuff out from doc-package and into content package
