+++ tools/api-builder/content-package/readers/content.js
@@ -1,9 +1,9 @@
 var path = require('canonical-path');
 
 /**
- * @dgService ngdocFileReader
+ * @dgService
  * @description
- * This file reader will pull the contents from a text file (by default .ngdoc)
+ * This file reader will pull the contents from a text file (by default .md)
  *
  * The doc will initially have the form:
  * ```
@@ -13,13 +13,13 @@ var path = require('canonical-path');
  * }
  * ```
  */
-module.exports = function ngdocFileReader() {
-  var reader = {
-    name: 'ngdocFileReader',
+module.exports = function contentFileReader() {
+  return {
+    name: 'contentFileReader',
     defaultPattern: /\.md$/,
     getDocs: function(fileInfo) {
 
-      // We return a single element array because ngdoc files only contain one document
+      // We return a single element array because content files only contain one document
       return [{
         docType: 'guide',
         content: fileInfo.content,
@@ -27,6 +27,4 @@ module.exports = function ngdocFileReader() {
       }];
     }
   };
-
-  return reader;
 };
